+++ src/Commands.cpp
@@ -229,6 +229,7 @@ static SeqStrings gCommandNames =
     "CmdChangeBackgroundColor\0"
     "CmdSetTabColor\0"
     "CmdCompressPdf\0"
+    "CmdDecompressPdf\0"
     "CmdNone\0"
     "\0";
 
@@ -448,6 +449,7 @@ static i32 gCommandIds[] = {
     CmdChangeBackgroundColor,
     CmdSetTabColor,
     CmdCompressPdf,
+    CmdDecompressPdf,
     CmdNone,
 };
 
@@ -667,6 +669,7 @@ SeqStrings gCommandDescriptions =
     "Change Background Color\0"
     "Set Tab Color\0"
     "Compress PDF\0"
+    "Decompress PDF\0"
     "Do nothing\0"
     "\0";
 // clang-format on